LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/text-decoration-thickness-expected.txt


PASS Can set 'text-decoration-thickness' to CSS-wide keywords
FAIL Can set 'text-decoration-thickness' to var() references assert_equals: expected 2 but got 1
PASS Can set 'text-decoration-thickness' to the 'auto' keyword
FAIL Can set 'text-decoration-thickness' to a length assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSUnitValue]"
FAIL Can set 'text-decoration-thickness' to a percent Invalid values
PASS Setting 'text-decoration-thickness' to a time throws TypeError
PASS Setting 'text-decoration-thickness' to an angle throws TypeError
PASS Setting 'text-decoration-thickness' to a flexible length throws TypeError
FAIL Setting 'text-decoration-thickness' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'text-decoration-thickness' to a URL throws TypeError
PASS Setting 'text-decoration-thickness' to a transform throws TypeError
